Consolidate DoReturnOp and NativeFunction return logic
Now share the same return from start of return valuestack location. Natives must declare their return count, default is 1.

+++ ulox/ulox.core/Package/Runtime/Types/Chunk.cs
@@ -31,8 +31,8 @@ namespace ULox
         public FunctionType FunctionType { get; internal set; }
         public bool IsLocal => FunctionType == FunctionType.LocalFunction || FunctionType == FunctionType.LocalMethod;
         public bool IsPure => FunctionType == FunctionType.PureFunction;
-        public int Arity => ArgumentConstantIds.Count;
-        public int ReturnCount => ReturnConstantIds.Count;
+        public byte Arity => (byte)ArgumentConstantIds.Count;
+        public byte ReturnCount => (byte)ReturnConstantIds.Count;
         public int UpvalueCount { get; internal set; }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
